syracuse-tools: move the project to the common root
