--- syracuse-tools/.bzrignore
@@ -1,8 +1,0 @@
-~*
-*.log
-*.suo
-*.user
-bin/
-obj/
-./TestResults/
-./Hosting/
