--- syracuse-tools/Syracuse Tools.sln
@@ -1,26 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 10
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syracuse.UnitTesting", "Syracuse.UnitTesting\Syracuse.UnitTesting.csproj", "{20FF4601-1283-4755-91C2-C3028C1299FA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syracuse.Common", "Syracuse.Common\Syracuse.Common.csproj", "{F9105A50-AE13-4A78-81C0-2DDCD593A6A4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{20FF4601-1283-4755-91C2-C3028C1299FA}.Debug|x86.ActiveCfg = Debug|x86
-		{20FF4601-1283-4755-91C2-C3028C1299FA}.Debug|x86.Build.0 = Debug|x86
-		{20FF4601-1283-4755-91C2-C3028C1299FA}.Release|x86.ActiveCfg = Release|x86
-		{20FF4601-1283-4755-91C2-C3028C1299FA}.Release|x86.Build.0 = Release|x86
-		{F9105A50-AE13-4A78-81C0-2DDCD593A6A4}.Debug|x86.ActiveCfg = Debug|x86
-		{F9105A50-AE13-4A78-81C0-2DDCD593A6A4}.Debug|x86.Build.0 = Debug|x86
-		{F9105A50-AE13-4A78-81C0-2DDCD593A6A4}.Release|x86.ActiveCfg = Release|x86
-		{F9105A50-AE13-4A78-81C0-2DDCD593A6A4}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
